Lang_approach/requirements.txt
PyPDF2 
langchain 
google-generativeai 
python-dotenv 
faiss-cpu
pytorch